rdagent/scenarios/rl/autorl_bench/tasks/alfworld/requirements.txt
# ALFWorld 评测依赖
alfworld>=0.4.0
openai  # API 模式
pyyaml
numpy
